refactor(builds-optimizer): builds-optimizer folder in usescases and domain
