+++ src/adapters/secondaries/characters-stats/character-stats-type.ts
@@ -1,7 +1,7 @@
-import { CharacterStatsPerLevel } from '../../../domain/models/character-statistics';
-import { ExistingCharacters } from '../../../domain/models/character';
-import { Levels } from '../../../domain/models/levels';
-import { WeaponType } from '../../../domain/models/weapon';
+import { CharacterStatsPerLevel } from '../../../domain/builds-optimizer/models/character-statistics';
+import { ExistingCharacters } from '../../../domain/builds-optimizer/models/character';
+import { Levels } from '../../../domain/builds-optimizer/models/levels';
+import { WeaponType } from '../../../domain/builds-optimizer/models/weapon';
 
 export interface CharacterWithStats {
   name: ExistingCharacters;
