+++ src/adapters/secondaries/weapons-stats/weapon-stats-type.ts
@@ -1,5 +1,5 @@
-import { WeaponType } from '../../../domain/models/weapon';
-import { Levels } from '../../../domain/models/levels';
+import { WeaponType } from '../../../domain/builds-optimizer/models/weapon';
+import { Levels } from '../../../domain/builds-optimizer/models/levels';
 
 export type WeaponLevels = Record<Levels, { atk: number; bonusStat?: { [bonusStat: string]: number } }>;
