+++ src/di/characters-di.ts
@@ -1,5 +1,5 @@
 import { InMemoryCharactersRepository } from '../adapters/secondaries/in-memory-characters-repository';
-import { CharactersHandler } from '../usescases/characters-handler';
+import { CharactersHandler } from '../usescases/builds-optimizer/characters-handler';
 
 export const CharactersDI = {
   charactersHandler: new CharactersHandler(new InMemoryCharactersRepository()),
