+++ src/di/weapons-di.ts
@@ -1,5 +1,5 @@
 import { InMemoryWeaponsRepository } from '../adapters/secondaries/in-memory-weapons-repository';
-import { WeaponsHandler } from '../usescases/weapons-handler';
+import { WeaponsHandler } from '../usescases/builds-optimizer/weapons-handler';
 
 export const WeaponsDI = {
   weaponsHandler: new WeaponsHandler(new InMemoryWeaponsRepository()),
