+++ src/domain/builds-optimizer/models/character.ts
