+++ src/domain/builds-optimizer/models/weapon.ts
