+++ src/usescases/builds-optimizer/builds-forms-handler.ts
@@ -1,6 +1,6 @@
-import { WeaponLevels } from '../adapters/secondaries/weapons-stats/weapon-stats-type';
-import { Levels } from '../domain/models/levels';
-import { WeaponsRepository } from '../domain/weapons-repository';
+import { WeaponLevels } from '../../adapters/secondaries/weapons-stats/weapon-stats-type';
+import { Levels } from '../../domain/builds-optimizer/models/levels';
+import { WeaponsRepository } from '../../domain/builds-optimizer/weapons-repository';
 
 export interface SelectOption {
   value: string | number;
